+++ CHANGELOG.md
@@ -2,9 +2,9 @@
 
 ## Next
 
-### Fixed
-
 - Fixed invalid lexical graph relationships causing "Relationship references unknown start node" errors during parquet import when nodes are pruned.
+- Make rate limit handler open to which exceptions it can retry on
+- Fix the initialization of the vertexai LLM class
 
 ## 1.13.0